requirements.txt
# Core dependencies
anthropic>=0.40.0
claude-agent-sdk>=0.1.0
python-dotenv>=1.0.0

# Optional: OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Optional: Langfuse for LLM observability
langfuse>=2.0.0

# Optional: Visualization
matplotlib>=3.7.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
